pyproject.toml
[tool.poetry]
name = "webtrh"
version = "0.1.0"
description = ""
authors = ["Radim Mifka <radimmifka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
lxml = "^4.6.3"
beautifulsoup4 = "argparse"
argparse = "^1.4.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
lxml = "^4.6.3"
argparse = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
